requirements.txt
Flask==3.1.0
gunicorn==21.2.0
SQLAlchemy==2.0.39
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.2
Werkzeug==3.1.3
python-dotenv==1.1.0
psycopg2-binary==2.9.9